Harden ProductLobby for launch: type-safety, security, core lobby flow, UX
Toward go-to-market (10 populated, actively-engaged campaigns):

- TypeScript: fix ~800 schema-drift errors across ~245 files → 0; re-enable
  ignoreBuildErrors: false. Production build passes with type-checking on.
- Security (audits in 04-Development/audits/): fix bank-detail IDOR, JSON-LD
  stored XSS, magic-link direct-mode account takeover, always-pass brand
  verification stubs, unguarded DB-wipe endpoint; tighten middleware route
  matching; 30-minute magic-link expiry.
- Fake data: delete 14 gimmick features (routes + orphaned components), gate
  14 partials behind ENABLED_FEATURES flag (src/lib/feature-flags.ts).
- Core flow: fix lobby creation (500'd on wrong Prisma relations) and an
  app-wide modal-centering bug (slide-up keyframe clobbered translate).
- UX: real per-campaign preferences (was hardcoded Nike demo), persist lobby
  across sign-in, honour ?redirect through the email magic-link flow, hide
  unconfigured OAuth buttons, brand-neutral empty states, real homepage/
  platform stats, dedupe page titles, seed brands.
- Tests: 55-test Playwright persona suite (e2e/); visitor suite green.

NOTE: the CI fix (remove continue-on-error from .github/workflows/ci.yml so
lint+typecheck actually block) is intentionally NOT in this commit — pushing
workflow changes needs an OAuth token with `workflow` scope. Apply separately.

+++ src/components/campaigns/goal-tracker.tsx
@@ -94,7 +94,7 @@ const calculateDaysRemaining = (deadline: string): number => {
 const formatDate = (dateString: string): string => {
   const date = new Date(dateString)
   return date.toLocaleDateString('en-GB', {
-    day: 'short',
+    day: 'numeric',
     month: 'short',
     year: 'numeric',
   })
